re-organized project structure, add .circleci/config.yml
