+++ requirements.txt
@@ -1,0 +1,3 @@
+pydicom==1.0.2
+setuptools==39.2.0
+extractCMRRPhysio==0.1
